Create PythonistaClockWidget.py
A "gadget" - like panel running down the right side of the Pythonista app which currently features an ultra smooth animated SVG analog clock and the ability to speak the current time and location on demand.  The significant aspects of the app is that it is a web view that hosts html code that itself embeds the svg clock and the javascript that manipulates it. So the relevant technologies are that a  python-based web view ui control runs in a sandboxed ios app (Pythonista) on an idevice in which the web view hosts html that embeds svg and javascript to do the work of an ultra-smooth second hand of a realistic analog clock rendered without additional style sheet tricks or image files.